+++ src/main/java/couponModel/Cprice.java
@@ -35,8 +35,8 @@ public class Cprice extends HttpServlet{
 		Coupon cc=client.target("http://localhost:8080/coupon-service/webapi/myresource/get").queryParam("var",Iitem).request().get(Coupon.class);
 		Coupon cc2=client.target("http://localhost:8080/coupon-service/webapi/myresource/get").queryParam("var",Iitem2).request().get(Coupon.class);
 		
-		int totalprice= Ipitem1*cc.Discount + Ipitem2*cc2.Discount;
-		String Stp= Integer.toString(totalprice);
+		Double totalprice= Ipitem1*cc.Discount + Ipitem2*cc2.Discount;
+		String Stp= Double.toString(totalprice);
 		out.println("hellloooooo");
 		//	request.setAttribute("tp", Stp);
 	//	request.getRequestDispatcher("Cart1.jsp").forward(request, response);
